requirements.txt
python-telegram-bot
python-dotenv
pydantic
vk-api
